// frontend/src/app/models/financial-management/financePaymentData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface FinancePaymentData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** บัญชีเงินฝากที่รับเงิน */
    bankAccountId?: number;
    /** ประเภทรายการ/บัตร */
    cardType: number;
    /** chargeId */
    chargeId?: string;
    /** custom courseCode */
    courseCode?: string;
    /** courseId */
    courseId?: number;
    /** fk coursepublic_main = coursepublic_id */
    coursepublicId?: number;
    /** fk อ้างอิงการนำเข้า finance_import_detail.detail_id */
    detailImpId?: number;
    first?: number;
    /** fk member_course  = member_course_id */
    memberCourseId?: number;
    /** custom ชื่อ(อังกฤษ) */
    memberFirstnameEn?: string;
    /** custom ชื่อ(ภาษาไทย) */
    memberFirstnameTh?: string;
    /** fk mas_member.member_id */
    memberId?: number;
    /** custom ชื่อ(อังกฤษ) */
    memberLastnameEn?: string;
    /** custom ชื่อ(ภาษาไทย) */
    memberLastnameTh?: string;
    /** memberReceiptViewFlag */
    memberReceiptViewFlag?: boolean;
    mode?: string;
    /** จำนวนเงิน */
    paymentAmount?: number;
    /** PK */
    paymentId?: number;
    /** สถานะการชำระเงิน */
    paymentStatus?: number;
    /** ประเภทรูปแบบการชำระเงิน */
    paymentType: number;
    /** custom ประเภทรูปแบบการชำระเงิน (อังกฤษ) */
    paymentTypeEn?: string;
    /** custom ประเภทรูปแบบการชำระเงิน (ภาษาไทย) */
    paymentTypeTh?: string;
    /** custom ชื่อหลักสูตร(อังกฤษ) */
    publicNameEn?: string;
    /** custom ชื่อหลักสูตร(ภาษาไทย) */
    publicNameTh?: string;
    /** Path file copy ใบเสร็จ ที่เข้ารหัส CA แล้ว */
    receiptCopyCaPath?: string;
    /** Path file copy ใบเสร็จ ที่ยังไม่เข้ารหัส CA */
    receiptCopyNonCaPath?: string;
    /** วันที่ชำระเงิน */
    receiptDate?: string;
    /** custom วันที่ชำระเงินใช้ในการค้นหา */
    receiptDateList?: string[];
    /** เลขที่ใบเสร็จ */
    receiptNo?: string;
    /** Path file original ใบเสร็จ ที่เข้ารหัส CA แล้ว */
    receiptOriginalCaPath?: string;
    /** Path file original ใบเสร็จ ที่ยังไม่เข้ารหัส CA */
    receiptOriginalNonCaPath?: string;
    /** Ref1 */
    ref1?: string;
    /** Ref2 */
    ref2?: string;
    rowNum?: number;
    size?: number;
    /** studyStatus */
    studyStatus?: number;
    /** Transaction Agent */
    transactionAgent?: string;
    /** Transaction Browser */
    transactionBrowser?: string;
    /** Transaction Datetime */
    transactionDatetime?: string;
    /** Transaction ID */
    transactionId?: string;
    /** Transaction IP */
    transactionIp?: string;
    /** Transaction OS */
    transactionOs?: string;
    /** Transaction Pay Account */
    transactionPayAccount?: string;
    /** Raw Data to Bank */
    transactionRawData?: string;
}